test(auth): merapihkan test login dan juga menambahkan beberapa test untuk refresh, login, logout
